Implement basic restacking support
The compositor now allows to restack surfaces into arbitrary locations in the
z-order.
As a first application, these functions are used to implement functions to
raise a window to the top or drop them to the bottom of the z-order, which
are used to fix the behaviour of floating windows. Windows now restack to
the top of the z-order whenever they become floating or when they are moved,
which is the same behaviour as dwm. When a window becomes tiled it gets
dropped to the bottom of the z-order, thus ensuring that floating windows are
always above tiled windows.